removed superfluous conversion

+++ contracts/ConditionalTokensMany.sol
@@ -207,7 +207,7 @@ contract ConditionalTokensMany is ERC1155 {
     function _collateralBalanceOf(IERC20 collateralToken, uint64 market, uint64 outcome, address customer) internal view
         returns (uint256 conditonalBalance, uint256 collateralBalance)
     {
-        uint256 numerator = uint256(payoutNumerators[outcome][customer]);
+        uint256 numerator = payoutNumerators[outcome][customer];
         uint256 denominator = payoutDenominator[outcome];
         conditonalBalance = balanceOf(customer, _conditionalTokenId(market, customer));
         uint256 collateralTotalBalance = collateralTotals[address(collateralToken)][market][outcome];
